+++ app/controllers/dashboard_controller.rb
@@ -1,7 +1,9 @@
 class DashboardController < ApplicationController
+  breadcrumb "Dashboard", :dashboard_path
+
   def show
-    @current_user = current_user
-    @journeys = Journey.not_remove.includes(:category).where(user_id: current_user.id)
+    journeys = Journey.not_remove.includes(:category).where(user_id: current_user.id)
+    @journeys = journeys.map { |j| JourneyPresenter.new(j) }
   end
 
   # :nocov: