+++ app/controllers/users_controller.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+class UsersController < ApplicationController
+  def index
+    render :show
+  end
+
+  def show; end
+end