+++ app/models/support/case.rb
@@ -1,10 +1,9 @@
 # frozen_string_literal: true
 
-# The top-level entity tracking cases
-# Support::Case has one Support::Enquiry
-# Support::Case belongs_to Support::Category
-
 module Support
+  #
+  # A case is opened from a "support enquiry" dealing with a "category of spend"
+  #
   class Case < ApplicationRecord
     include Support::Documentable
 
@@ -12,25 +11,34 @@ module Support
     belongs_to :category, class_name: "Support::Category"
     belongs_to :agent, class_name: "Support::Agent", optional: true
 
-    # User-defined support levels
+    # Support level
     #
-    # L1       - Advice and guidance only
-    # L2       - Pointing to a framework / catalogue
-    # L3       - Helping school run a framework but school doing system work
-    # L4       - Run framework on behalf of school
-    # L5       - Run bespoke procurement
+    #   L1       - Advice and guidance only
+    #   L2       - Pointing to a framework / catalogue
+    #   L3       - Helping school run a framework but school doing system work
+    #   L4       - Run framework on behalf of school
+    #   L5       - Run bespoke procurement
     enum support_level: { L1: 0, L2: 1, L3: 2, L4: 3, L5: 4 }
 
-    # User-defined state
+    # State
     #
-    # initial (default)
-    # open
-    # resolved
-    # pending
-    # closed
-    # pipeline
-    # no_response
+    #   initial (default)
+    #   open
+    #   resolved
+    #   pending
+    #   closed
+    #   pipeline
+    #   no_response
     enum state: { initial: 0, open: 1, resolved: 2, pending: 3, closed: 4, pipeline: 5, no_response: 6 }
+
+    before_validation :generate_ref
+    validates :ref, uniqueness: true, length: { is: 6 }, format: { with: /\A\d+\z/, message: "numbers only" }
+
+    # Called before validation to assign 6 digit incremental number (from last case or the default 000000)
+    # @return [String]
+    def generate_ref
+      self.ref = (Support::Case.last&.ref || sprintf("%06d", 0)).next
+    end
 
     # TODO: Replace with ActiveRecord association
     def interactions