+++ db/schema.rb
@@ -39,13 +39,14 @@ ActiveRecord::Schema.define(version: 2021_09_13_154900) do
 
   create_table "categories", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "title", null: false
-    t.string "description"
+    t.string "description", null: false
     t.string "contentful_id", null: false
     t.jsonb "liquid_template", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.datetime "created_at", precision: 6, default: -> { "CURRENT_TIMESTAMP" }, null: false
+    t.datetime "updated_at", precision: 6, default: -> { "CURRENT_TIMESTAMP" }, null: false
     t.integer "journeys_count"
-    t.string "slug"
+    t.string "slug", null: false
+    t.index ["contentful_id"], name: "index_categories_on_contentful_id", unique: true
   end
 
   create_table "checkbox_answers", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
@@ -166,21 +167,25 @@ ActiveRecord::Schema.define(version: 2021_09_13_154900) do
     t.string "request_text"
     t.integer "support_level"
     t.integer "status"
-    t.integer "state"
+    t.integer "state", default: 0
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
     t.uuid "agent_id"
     t.index ["category_id"], name: "index_support_cases_on_category_id"
-    t.index ["ref"], name: "index_support_cases_on_ref"
+    t.index ["ref"], name: "index_support_cases_on_ref", unique: true
     t.index ["state"], name: "index_support_cases_on_state"
     t.index ["status"], name: "index_support_cases_on_status"
     t.index ["support_level"], name: "index_support_cases_on_support_level"
   end
 
   create_table "support_categories", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
-    t.string "name", null: false
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
+    t.string "title", null: false
+    t.datetime "created_at", precision: 6, default: -> { "CURRENT_TIMESTAMP" }, null: false
+    t.datetime "updated_at", precision: 6, default: -> { "CURRENT_TIMESTAMP" }, null: false
+    t.string "slug"
+    t.string "description"
+    t.index ["slug"], name: "index_support_categories_on_slug", unique: true
+    t.index ["title"], name: "index_support_categories_on_title", unique: true
   end
 
   create_table "support_documents", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
@@ -208,6 +213,32 @@ ActiveRecord::Schema.define(version: 2021_09_13_154900) do
     t.datetime "updated_at", precision: 6, null: false
     t.index ["case_id"], name: "index_support_enquiries_on_case_id"
     t.index ["support_request_id"], name: "index_support_enquiries_on_support_request_id"
+  end
+
+  create_table "support_requests", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.string "user_id"
+    t.uuid "journey_id"
+    t.uuid "category_id"
+    t.string "message"
+    t.string "school_name"
+    t.string "school_urn"
+    t.datetime "created_at", precision: 6, null: false
+    t.datetime "updated_at", precision: 6, null: false
+    t.string "phone_number"
+    t.index ["category_id"], name: "index_support_requests_on_category_id"
+    t.index ["journey_id"], name: "index_support_requests_on_journey_id"
+    t.index ["user_id"], name: "index_support_requests_on_user_id"
+  end
+
+  create_table "support_sub_categories", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
+    t.uuid "category_id", null: false
+    t.string "title", null: false
+    t.string "description"
+    t.string "slug"
+    t.datetime "created_at", precision: 6, default: -> { "CURRENT_TIMESTAMP" }, null: false
+    t.datetime "updated_at", precision: 6, default: -> { "CURRENT_TIMESTAMP" }, null: false
+    t.index ["slug"], name: "index_support_sub_categories_on_slug", unique: true
+    t.index ["title"], name: "index_support_sub_categories_on_title", unique: true
   end
 
   create_table "tasks", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|