+++ spec/factories/category.rb
@@ -2,12 +2,9 @@ FactoryBot.define do
   factory :category do
     title { "category title" }
     description { "category description" }
-    contentful_id { "12345678" }
+    sequence(:contentful_id) { |n| n }
     liquid_template { "Your answer was {{ answer_47EI2X2T5EDTpJX9WjRR9p }}" }
-    # Tests covering data migration in production use factories
-    if ENV["POST_MIGRATION_CHANGES"] == "true"
-      slug { "slug" }
-    end
+    slug { "slug" }
 
     trait :catering do
       title { "Catering" }