+++ spec/factories/support/category.rb
@@ -1,5 +1,9 @@
 FactoryBot.define do
   factory :support_category, class: "Support::Category" do
-    name { "support category title" }
+    sequence(:title) { |n| "support category title #{n}" }
+    sequence(:slug) { |n| "support category slug #{n}" }
+    description { "support category description" }
+
+    # association :category, factory: :support_sub_category
   end
 end