+++ spec/features/self-serve/journeys/category_spec.rb
@@ -19,6 +19,10 @@ RSpec.feature "A journey page has" do
   context "when the category is 'catering'" do
     let(:fixture) { "radio-question" }
 
+    specify "breadcrumbs" do
+      expect(page).to have_breadcrumbs ["Dashboard", "Create specification"]
+    end
+
     specify do
       expect(page).to have_a_journey_path
     end
@@ -35,6 +39,10 @@ RSpec.feature "A journey page has" do
 
   context "when the category is 'MFDs'" do
     let(:fixture) { "mfd-radio-question" }
+
+    specify "breadcrumbs" do
+      expect(page).to have_breadcrumbs ["Dashboard", "Create specification"]
+    end
 
     specify do
       expect(page).to have_a_journey_path