+++ spec/models/support/case_spec.rb
@@ -3,7 +3,7 @@ RSpec.describe Support::Case, type: :model do
 
   it "belongs to a category" do
     expect(support_case.category).not_to be_nil
-    expect(support_case.category.name).to eql "support category title"
+    expect(support_case.category.title).to match /support category title \d/
   end
 
   context "with documents" do
@@ -15,6 +15,26 @@ RSpec.describe Support::Case, type: :model do
     end
   end
 
-  it { is_expected.to define_enum_for(:support_level).with(%i[L1 L2 L3 L4 L5]) }
-  it { is_expected.to define_enum_for(:state).with(%i[initial open resolved pending closed pipeline no_response]) }
+  it { is_expected.to define_enum_for(:support_level).with_values(%i[L1 L2 L3 L4 L5]) }
+  it { is_expected.to define_enum_for(:state).with_values(%i[initial open resolved pending closed pipeline no_response]) }
+
+  describe "#generate_ref" do
+    context "when no cases exist" do
+      it "generates a reference starting at 1" do
+        expect(support_case.ref).to eql "000001"
+      end
+    end
+
+    context "when cases already exist" do
+      before do
+        create_list(:support_case, 10)
+      end
+
+      let!(:new_case) { create(:support_case) }
+
+      it "generates an incrementing reference" do
+        expect(new_case.ref).to eql "000011"
+      end
+    end
+  end
 end