+++ spec/support/contentful_helpers.rb
@@ -146,8 +146,10 @@ module ContentfulHelpers
       specification_template: hash_response.dig("fields", "specificationTemplate"),
       specification_template_part2: hash_response.dig("fields", "specificationTemplatePart2"),
       combined_specification_template: combined_specification_template,
-      environment: double(id: "test"),
+      environment: double(id: hash_response.dig("sys", "environment", "sys", "id")),
+      space: double(id: hash_response.dig("sys", "space", "sys", "id")),
       slug: hash_response.dig("fields", "slug"),
+      updated_at: Time.zone.parse(hash_response.dig("sys", "updatedAt")),
     )
 
     allow(category_double).to receive(:combined_specification_template=)